Rename staging Heroku module to sandbox
Renames the staging TF module name to `api_sandbox`.

Note - this change requires significant changes to the Terraform state.
Most resources inside the `api_staging` module simply need to be moved
to the `api_sandbox` module; all of those changes are encapsulated in
`moved.tf`.
However, the resources that we actually want to change,
such as the route53 record and heroku domain config, need to be
deleted and recreated. For stability purposes, we do not want to kill
`api-staging.dandiarchive.org` until we are ready to fully switch over
to the `sandbox` domain; to facilitate this, I've pulled those resources
out into the root module (see the new resources in
`staging_pipeline.tf`). Once we're ready to fully switch, we can simply
remove those resources.